update: Add maven dependecy and implemented null virefiers

+++ src/main/java/npi/autoestudo/todoapp/service/TodoService.java
@@ -5,19 +5,20 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import npi.autoestudo.todoapp.model.Todo;
-import npi.autoestudo.todoapp.repository.TodoRepository;;
+import npi.autoestudo.todoapp.repository.TodoRepository;
 
 @Service
 public class TodoService {
 
-    private TodoRepository todoRepository;
+    private final TodoRepository todoRepository;
 
     public TodoService(TodoRepository todoRepository) {
         this.todoRepository = todoRepository;
     }
 
     public List<Todo> create(Todo todo) {
-        todoRepository.save(todo);
+        if (todo != null)
+            todoRepository.save(todo);
         return list();
     }
 
@@ -26,12 +27,14 @@ public class TodoService {
     }
 
     public List<Todo> update(Todo todo) {
-        todoRepository.save(todo);
+        if (todo != null)
+            todoRepository.save(todo);
         return list();
     }
 
     public List<Todo> delete(Long id) {
-        todoRepository.deleteById(id);
+        if (id != null)
+            todoRepository.deleteById(id);
         return list();
     }
 }
